v1.1
Moved the calculation of the distance matrix to a separate python file. The reason for doing this was the computational intensity of computing N(N-1)/2 numerical integrations for the distance, rather than N(N-1)/2 distance formula calculations as before. The terrain functionality has been fully implemented - the distance is now the arc length along the terrain surface. Circulation around a central Gaussian mountain has been observed in the results of netsim.